hook up the beneficial owners step
