+++ src/libs/Navigation/linkingConfig.js
@@ -129,11 +129,6 @@ export default {
                     NewWorkspace_Root: ROUTES.WORKSPACE_NEW,
                 },
             },
-            BeneficialOwners: {
-                screens: {
-                    BeneficialOwners_Root: 'beneficial',
-                },
-            },
         },
     },
 };
