--- src/pages/ReimbursementAccount/ACHContractStep.js
@@ -1,5 +1,0 @@
-import React from 'react';
-import {View} from 'react-native';
-
-const ACHContractStep = () => <View />;
-export default ACHContractStep;
